refactor(m6): split embedding and clustering into independent pipelines
- NEW: gaia/lkm/pipelines/embedding.py — standalone embedding pipeline
- RENAME: discovery.py → cluster.py — clustering only, no embedding
- UPDATE: scripts/run_full_discovery.py — calls embedding then cluster
  with --embedding-only and --cluster-only flags

Pipelines are now independently runnable:
  python -m gaia.lkm.pipelines.embedding        # just compute embeddings
  python -m gaia.lkm.pipelines.cluster           # just cluster
  python scripts/run_full_discovery.py            # both
